Refactor yxf to enable use as a library.
- Split __main__.py into modules:
  - yxf/excel.py: Excel/XLSForm reading and writing
  - yxf/yaml.py: YAML conversion (string-based)
  - yxf/markdown.py: Markdown conversion (string-based)
  - yxf/cli.py: CLI logic and file I/O

- Library functions now accept IO objects (Excel: BinaryIO) or strings
  (YAML/Markdown) instead of filenames for programmatic use

- Exposed public API through yxf/__init__.py for external projects
